Ultima version

+++ Scripts OLAP/procedimientos/rent_stats_by_duration.sql
@@ -1,12 +1,12 @@
 -- Dar el número de alquileres y el monto cobrado, por duración del préstamo
 
 CREATE OR REPLACE FUNCTION rent_stats_by_duration ()
-RETURNS TABLE (days smallint, rent_count INT, amount_charged NUMERIC(5,2))
+RETURNS TABLE (days smallint, rent_count BIGINT, amount_charged NUMERIC(5,2))
 LANGUAGE plpgsql    
 AS $$
 BEGIN
     RETURN QUERY
-        SELECT D.day_count, SUM(RS.rental_count)::INT, SUM(RS.amount_sold)
+        SELECT D.day_count, SUM(RS.rental_count), SUM(RS.amount_sold)
         FROM rental_stats RS
         INNER JOIN duration D ON D.duration_id = RS.duration_id
         GROUP BY D.day_count
